+++ src/wl_wordlist.py
@@ -393,7 +393,7 @@ class Wl_Worker_Wordlist(wl_threading.Wl_Worker):
             texts = []
 
             settings = self.main.settings_custom['wordlist']
-            files = self.main.wl_files.get_selected_files()
+            files = list(self.main.wl_file_area.get_selected_files())
 
             # Frequency
             for file in files:
@@ -622,7 +622,7 @@ def generate_table(main, table):
             wl_msgs.wl_msg_fatal_error(main)
 
     settings = main.settings_custom['wordlist']
-    files = main.wl_files.get_selected_files()
+    files = list(main.wl_file_area.get_selected_files())
 
     if wl_checking_files.check_files_on_loading(main, files):
         dialog_progress = wl_dialogs_misc.Wl_Dialog_Progress_Process_Data(main)
@@ -694,7 +694,7 @@ def generate_fig(main):
             wl_figs.show_fig()
 
     settings = main.settings_custom['wordlist']
-    files = main.wl_files.get_selected_files()
+    files = list(main.wl_file_area.get_selected_files())
 
     if wl_checking_files.check_files_on_loading(main, files):
         dialog_progress = wl_dialogs_misc.Wl_Dialog_Progress_Process_Data(main)